Move newer components to correct names, temp rename of old metadata component
