+++ app/templates/components/_workflow-metadata.hbs
@@ -3,7 +3,11 @@
   Form {{displayFormStep}} of {{schemas.length}}
 </h4>
 {{metadata-form
+  currentUser=currentUser
   schema=schema
+  submission=submission
+  doiInfo=doiInfo
+  currentFormStep=currentFormStep
   nextForm=(action "nextForm")
   previousForm=(action "previousForm")}}
 {{yield}}
