+++ app/templates/components/workflow-metadata.hbs
@@ -2,12 +2,10 @@
 <h4 class="font-weight-light pull-right" style="margin-top:6.5px;">
   Form {{displayFormStep}} of {{schemas.length}}
 </h4>
+{{#if currentSchema}}
 {{metadata-form
-  currentUser=currentUser
-  schema=schema
-  submission=submission
-  doiInfo=doiInfo
-  currentFormStep=currentFormStep
+  schema=currentSchema
   nextForm=(action "nextForm")
   previousForm=(action "previousForm")}}
+{{/if}}
 {{yield}}
